pstp: add ids to main divs.
By default, DocBook adds anchors with id attributes within the header.  This
produces invalid code.  The option to add id attributes to the relevant divs
instead is incompletely implemented and will result in duplicate ids and
therefore invalid code.  So we move the id attributes from the anchors onto the
divs, and the anchors are simply removed as before.