atom-pstp: convert span.indent to a tab.
In regular HTML and XHTML, span.indent appears as an indent of 2 ems, according
to the CSS stylesheets.  However, CSS is not referenced in Atom feeds, and using
the style attribute is not a solution since it is often stripped (for security
reasons).  Therefore, we replace those tags with an actual tab again, so that
they appear approximately as intended (such as in poems).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>