fo-poetry: make indents work in odd situations.
Because the indent code is only called for text blocks, if a line contains any
tags, the indent code will place an empty fo:inline element, which is removed.
We still want to indent in this case, so insert a U+200B (ZERO WIDTH SPACE) in
the fo:inline element if we have no other content.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>