Always produce the correct number of articles.
Previously, the stylesheet would produce no more than the given number of
articles, but would produce fewer articles if one or more articles were in
multiple categories.  That is, the stylesheet would count an article in multiple
categories as multiple articles.  This version should always produce the correct
number, at the expense of needing to use exsl:node-set().

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>