Force only the default namespace for XHTML.
Because Java-based tools automatically try to read and validate DTDs for XML
documents that have them, we have to ensure that only the default namespace is
present, since DTDs are generally not namespace-friendly.

Because the stylesheets need to address XHTML elements, which are in a
namespace, a prefix must be assigned to that namespace, since XSLT 1.0 treats
unprefixed element and attribute names as belonging to the null namespace, not
the default namespace.  However, this prefix must not escape the stylesheets;
otherwise, it causes the above-mentioned problems with DTDs.

Ironically, if an xsl:namespace-alias element is present, Xalan does the wrong
thing and emits the prefixed declaration for the namespace.  Therefore, we have
to remove those elements.  xsltproc does the right thing regardless, and Saxon
does the wrong thing in all cases (see Debian bug #568348).  Thus, we change
things for Xalan.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>