Add more fonts to the Free font lists.
This change still prefers the Nimbus fonts (which are PostScript Type 1) to the
new additions, the Freefont and Liberation families.  Note that the Liberation
Mono font is not considered for the Courier family, since although it is metric
compatible, its appearance (sans-serif) is a significant departure from the
PostScript Courier (serif).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>