fixup: use Roman numerals for book frontmatter.
The stylesheets normally use Arabic numbers for the titlepage and fly sheet of
books.  However, this causes a repetition of the numbering at the very least.
If there is any other frontmatter (prefaces, dedications, acknowledgements, and
the like), that frontmatter is numbered using Roman numerals, so there are
Arabic numbers, Roman numerals, and then Arabic numbers again.  Solve this by
always making the titlepage and fly sheet use Roman numerals to avoid this
problem.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>