Optionally expand verbatim tabs in to en spaces.
In some HTML/XHTML environments, such as ePUB, CSS is just not available or not
reliable.  This code allows the expansion of tabs in verbatim environments into
a set of en spaces so that they are properly expanded where CSS is not
available.  However, by default, the old behavior (of inserting a tab) is
enabled.  In either case, the class attribute is set; in CSS-free environments,
it will be ignored.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>