Work around bug in poppler's handling of empty titles.
Poppler treats an empty title for a PDF bookmark as a case where the item in
question is not a string; therefore, it does not display it or any of its
children (Debian bug #639614).  As a workaround, insert a single U+200B (ZERO
WIDTH SPACE) to trick libpoppler into displaying the bookmarks.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>