Add a stylesheet to fix up some Asciidoctor warts.
Asciidoctor formats verse blocks as a blockquote containing a literallayout.
This is great when the verse block is embedded in a piece of prose, but it's
suboptimal when formatting standalone poetry, as it results in the entire
piece being indented.  Fix this up by stripping out the blockquote in that
case.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>