xhtml5/base/cvt: use xhtml5 templates as base
When using the main XHTML templates as a base, XHTML 1.0 Transitional is
created, which results in a DTD being emitted. Since the DTDs are no
longer likely to be in catalogs on most systems, this results in build
failures. Use the XHTML5 templates as a base instead since they do not
use a DTD.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>