asciidoctor-fixup: fix quotations in sections, too
The main purpose of this template is to fix the extra indentation that
occurs from the blockquote when there is a single verse block in an
article, since that's probably a standalone poem and we don't need that
extra indentation.

However, this occurs as well for sections, which can occur when there's
a single poem with multiple sections, each of which contains a verse
block.  With Asciidoctor, however, there's no info block for sections,
so we need to handle a raw title as well.

This shouldn't affect other blocks because only elements that can be
top-level documents, various kinds of sections, and similar structures
can accept either the info or title elements, so we shouldn't
accidentally convert unrelated blockquotes that aren't verse blocks.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>