fo: avoid header and footer tables causing accessibility problems
When we generate PDFs with Apache FOP, it takes document structure from
the XSL-FO input and generates Tagged PDF with structural elements.
When it sees the header and footer tables, it emits table elements,
which are not really semantically interesting in the document, which
harms accessibility.

Since these aren't structurally part of the document and we don't want
them emitted in the tags, mark them with the `artifact` role, which
tells FOP to not emit them in the document structure.  As a result, our
document's accessibility is improved.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>